Seem to have fixed FPS drops when flying high up
Turns out Minecraft is constantly constructing entity instances,
checking if they can spawn, and spawning them if true. When you fly to
above roughly 180 on a normal world, the entities spawn, but you're more
than 128 blocks away so the game despawns them. The Rocket Squid
constructor registered the Rocket Squid instance with the Forge Event
Bus (needed to change the riders' rotation depending on the squids'
rotation), and it's being called something like every world tick. That
means there are thousands of Rocket Squids registering on the Event Bus
all the time, and to make matters worse they're not unregistering when
they die. This seems to be the cause of the FPS tanking at high y
levels. Now, registration only happens when the Rocket Squid is mounted,
and the Rocket Squid is unregistered on dismount. Also, when the Rocket
Squid dies while being ridden, it removes its passenger and therefore
unregisters itself. My FPS no longer drops to 0 when I fly high so this
seems to work/help.